Refactor: move RaftTypeConfig to separate file
