+++ openraft/src/raft/raft_inner.rs
@@ -11,7 +11,7 @@ use crate::core::raft_msg::external_command::ExternalCommand;
 use crate::core::raft_msg::RaftMsg;
 use crate::core::TickHandle;
 use crate::error::Fatal;
-use crate::raft::CoreState;
+use crate::raft::core_state::CoreState;
 use crate::storage::RaftLogStorage;
 use crate::AsyncRuntime;
 use crate::Config;
